WIP - Initial VAE prototype.
This is a composite of a number of commits.  See below for the changelog:

Removed version numbers from environment.yml.

Only keeping the python version number at this point.

Fixed an indent.

Added a base definition of a Model.

Expanded the Model API.

Added a VAE object stub.

Added (failing) tests for the VAE class.

Started adding in the actual useful "stuff".

Defined the encoder.

Implemented a decoder.

Encoder now outputs mean/stddev.

Created a torch module for the VAE.

This keeps the model separate from the training code, which makes the
code itself a bit cleaner.

Made changes to support the actual VAE model.

Added test for low-dimension projection.

Mu is now restricted using a hard-tanh.

The mean is going to be restricted to [-4,4] so that it can be just
slighly outside of the usual 3-sigma range for a normal distribution.

Added extra input check.

Discovered lack of input sanitization when writing the unit tests.

Added something to calculate model costs.

Added some test parameterization.

Added test for squeezed inputs into decoder.

Finished implementing the top-level VAE.

Got everything but training working.

Pretty much have everything working minus the training.  That seems to
still have some type errors that need to be resolved.

Fixed minor bug with batch processing.

Started on a Dataset to load image data.

This is mainly to make it possible to load image data for the model
training.

Fixed PNG load issue in the IO utils.

PNGs don't have any EXIF data, so the current verion of imread() will
fail.  This commit fixes that issue by ensuring the EXIF handling
doesn't happen with PNG images.

Finalized the Dataset class (mostly).